services/market-streamer/requirements.txt
aiohttp==3.9.5
google-cloud-pubsub==2.23.1
google-cloud-secret-manager==2.20.2
orjson==3.10.7
pydantic==2.8.2
